+++ types/executable/executable.go
@@ -459,19 +459,3 @@ func NewExecutableID(workspace, namespace, name string) string {
 	}
 	return fmt.Sprintf("%s/%s:%s", workspace, namespace, name)
 }
-
-func (e *ParallelRefConfig) RecordAttempt() {
-	e.retryAttempts++
-}
-
-func (e *ParallelRefConfig) AttemptedMaxTimes() bool {
-	return e.retryAttempts >= e.Retries+1
-}
-
-func (e *SerialRefConfig) RecordAttempt() {
-	e.retryAttempts++
-}
-
-func (e *SerialRefConfig) AttemptedMaxTimes() bool {
-	return e.retryAttempts >= e.Retries+1
-}
